% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunc.R
\name{betaBinom}
\alias{betaBinom}
\title{Beta binomial likelihood function alt formulation}
\usage{
betaBinom(k, n, p, w)
}
\arguments{
\item{k}{Observed data point}

\item{n}{Simulated data point}

\item{p}{fraction of population}

\item{w}{overdisperion parameter}
}
\value{
log likelihood
}
\description{
Beta binomial likelihood function alt formulation
}
