% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFuncs.R
\name{boonahDatFunc}
\alias{boonahDatFunc}
\title{batDataFunc}
\usage{
boonahDatFunc(ret = "all", species = "BFF", cutoff = 1636, fileLoc)
}
\arguments{
\item{ret}{which data type to return all or observed}

\item{species}{which species to return for BFF GHFF or LRFF}

\item{cutoff}{the cutoff value to use for serology}

\item{fileLoc}{location of raw data file}
}
\value{
database
}
\description{
batDataFunc
}
