% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodFunc.R
\name{likelihoodFuncBoonahStoch}
\alias{likelihoodFuncBoonahStoch}
\title{likelihood function}
\usage{
likelihoodFuncBoonahStoch(
  parms,
  N,
  k_PpSp,
  k_PnSp,
  k_PpSn,
  k_PnSn,
  S,
  I,
  Ia,
  E,
  R,
  PCR,
  time,
  assump = "EIR",
  urPredDat
)
}
\arguments{
\item{parms}{parameter values}

\item{N}{observed total population}

\item{I}{total infected}

\item{Ia}{infected adult males and adult females}

\item{E}{total exposed/latent}

\item{R}{total recovered}

\item{PCR}{observed PCR positive counts - out of N}

\item{time}{week of the year 1-52 for seasonal information}

\item{urPredDat}{predicted values for under roost from GAM}

\item{k}{observed sero positive}

\item{SN}{total susceptable}
}
\value{
log-likelihood value
}
\description{
likelihood function
}
