% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModCont.R
\name{modPlotFuncContinuous}
\alias{modPlotFuncContinuous}
\title{plot model from mcmc results}
\usage{
modPlotFuncContinuous(
  mcmc,
  assum = "EI",
  fxd = F,
  iters = 100,
  pName,
  birthType,
  Rt = F
)
}
\arguments{
\item{mcmc}{mcmc results}
}
\value{
figures
}
\description{
plot model from mcmc results
}
