% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModFunc.R
\name{modPlotFuncOLD}
\alias{modPlotFuncOLD}
\title{plot model from mcmc results}
\usage{
modPlotFuncOLD(mcmc, assum = "EI", fxd = F, osc = F)
}
\arguments{
\item{mcmc}{mcmc results}
}
\value{
figures
}
\description{
plot model from mcmc results
}
