% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModRuns.R
\name{modPlotFuncX}
\alias{modPlotFuncX}
\title{plot model from mcmc results}
\usage{
modPlotFuncX(
  mcmc,
  assum = "EI",
  fxd = F,
  iters = 100,
  pName,
  birthType,
  Rt = F,
  saveLoc
)
}
\arguments{
\item{mcmc}{mcmc results}
}
\value{
figures
}
\description{
plot model from mcmc results
}
