% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelStepFunc.R
\name{modStepJI}
\alias{modStepJI}
\title{model step function - runs model in steps using Odin, returning the model state at a subsequent time point}
\usage{
modStepJI(particles, mod, birthType, startTime, endTime)
}
\arguments{
\item{weightInput}{string of values for model state, time to run model from/to and model parameters}
}
\value{
model state at next timestep
}
\description{
model step function - runs model in steps using Odin, returning the model state at a subsequent time point
}
