% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metHast.R
\name{multiv.proposer}
\alias{multiv.proposer}
\title{function for adaptive blocked proposals based on var-covar matrix}
\usage{
multiv.proposer(current, sdProps)
}
\arguments{
\item{current}{current parameters}

\item{sdTune}{tuned s.d.}
}
\value{
proposal parameters
}
\description{
function for adaptive blocked proposals based on var-covar matrix
}
