% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterFunc.R
\name{paramsFunc}
\alias{paramsFunc}
\title{parameter function}
\usage{
paramsFunc(
  gamma_1_Val = 0,
  gamma_2_Val = 0,
  zeta_s = rnorm(1, 0.2, 0.1),
  sigma_2_Val = 0,
  mu_Val = 1.37,
  mj_Val = rnorm(1, 0.5, 0.1),
  m_Val = rnorm(1, 0.186, 0.05),
  omega_m_Val = rnorm(1, 0.8, 0.1),
  omega_2_Val = 0,
  epsilon_Val = 0,
  kappa_Val = rnorm(1, mean = 3.60206, sd = 0.1),
  c_Val = 2.113776,
  s_Val = rnorm(1, 130, 10),
  phi_Val = rnorm(1, 7.18, 0.1),
  rho_Val = 0,
  R0_Val = runif(1, 1, 25),
  Phi2_val = 1,
  oDist_s = 2,
  zeta_p = runif(1, 0.1, 1),
  pcrProb2 = runif(1, 0.1, 1),
  sigmaVer = ifelse(zeta_s > 0, 1, 2),
  gammaVer = ifelse(gamma_1_Val > 0, 1, 2),
  betaVer = 2,
  S2_val = runif(1, 1, 25),
  d_val = runif(1, 1, 10),
  oDist1 = 2,
  oDist_u = runif(1, 0, 3),
  c_val2 = 1,
  envOscType = 0,
  betaFX = 0,
  betaFXVal = 0
)
}
\arguments{
\item{x}{any parameter}
}
\value{
paramerter values
}
\description{
parameter function
}
