% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particleFilter.R
\name{pFilt}
\alias{pFilt}
\title{Particle filter function - runs model in steps between each observed data point
before using weighted re-sampling of model state at each step to start next}
\usage{
pFilt(
  n,
  iState,
  stepFun,
  obsData,
  prms,
  resM = F,
  cluster = F,
  assump,
  likelihoodFunc,
  full = F,
  mat = F,
  birthType
)
}
\arguments{
\item{n}{number of particles}

\item{iState}{function for calculating initial state of particles}

\item{stepFun}{function for taking parmeters and time point, running model in steps and outputting model state}

\item{obsData}{observed data to fit model to}

\item{prms}{model parameters}

\item{resM}{True/False whether to output likelihood values or results of simulation for plotting figures}

\item{cluster}{T/F if true, using the cluster if false, run locally}

\item{likeFunc}{likelihood function - obsolete}
}
\value{
if resM = F log likelihood value, if resM = T mean results of simulation
}
\description{
Particle filter function - runs model in steps between each observed data point
before using weighted re-sampling of model state at each step to start next
}
