% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metHast.R
\name{sequential.proposer}
\alias{sequential.proposer}
\title{Sequential proposal function}
\usage{
sequential.proposer(current, prmNum, sdProps)
}
\arguments{
\item{current}{current parameters}

\item{prmNum}{which parameter to update}

\item{sdTune}{s.d. tuning from tuner function}
}
\value{
proposal
}
\description{
Sequential proposal function
}
