% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metHast.R
\name{tuner}
\alias{tuner}
\title{proposal sd tuning function}
\usage{
tuner(curSd, acptR, curAcptR, maxSddProps)
}
\arguments{
\item{target}{acceptance ratio}

\item{current}{acceptance ratio}

\item{maximum}{proposal s.d.}
}
\value{
tuned s.d.
}
\description{
proposal sd tuning function
}
