% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metHast.R
\name{tunerSeq}
\alias{tunerSeq}
\title{sequentialTuner}
\usage{
tunerSeq(curSd, acptR, curAcptR, maxSddProps, i)
}
\arguments{
\item{i}{which parameter to work on}

\item{target}{acceptance ratio}

\item{current}{acceptance ratio}

\item{maximum}{proposal s.d.}
}
\value{
tuned s.d.
}
\description{
sequentialTuner
}
