% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFuncs.R
\name{weekification}
\alias{weekification}
\title{aggregate by week boonah data}
\usage{
weekification(obsDataBoonah)
}
\arguments{
\item{data}{boonah data}
}
\value{
boonah data aggregated by week
}
\description{
aggregate by week boonah data
}
