% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelStepFunc.R
\name{modStepJI}
\alias{modStepJI}
\title{model step function - runs model in steps using Odin, returning the model state at a subsequent time point}
\usage{
modStepJI(particles, mod, birthType, startTime, endTime)
}
\arguments{
\item{particles}{number of particles in run}

\item{mod}{redundant - needs removing}

\item{startTime}{current time in model}

\item{endTime}{time of next datapoint}

\item{birthtype}{maternal immunity on off}
}
\value{
simulated data for time step
}
\description{
model step function - runs model in steps using Odin, returning the model state at a subsequent time point
}
